tutorial/building-comfort/control-panel/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.6.1